requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0

# Google Gemini (PaLM) Integration
langchain-google-genai>=0.0.5
google-generativeai>=0.3.2

# Environment Variable Management
python-dotenv>=1.0.0

# Streamlit UI
streamlit>=1.28.0

# Pydantic for structured output parsing
pydantic>=1.10,<3.0

langchain-anthropic
# Typing support (built-in, no need to install separately)